bug: js async function get return value bug ❌
