+++ nodejs-files/stores/2022/11/27/sales.json
@@ -1,0 +1,3 @@
+{
+  "total": 58998.14
+}
